requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
alembic==1.13.3
requests==2.32.3
python-dotenv==1.0.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
gunicorn==21.2.0
